Home page
This will be the home page of this website. It will contan the following basic features;

1. Logo
The logo should be visible at the top of the website. It’s the core of the business’ branding and identity, after all. The logo is a tangible representation that encompasses the products or services and is a key piece for clients to recognize and connect with your brand. The logo on a website often doubles in functionality as a link to the homepage, so you want it to be in an obvious location within the header.
2. Navigation
Straightforward and intuitive navigation is another vital feature the website should include in the header. If the website is content-heavy, a search box might be a valuable asset to include here. The navigation menu should be easy to locate with items that make sense to new visitors. The navigation area is like a road map--it’s necessary to give visitors an overview of the content, as well as help them to locate the information they are looking for using the most efficient route. Descriptive and thorough navigation is also beneficial in reducing bounce rates.
3. Headline
Within a matter of seconds, the website needs to communicate to visitors what we have to offer. A headline with sub-headline or paragraph text should provide a clear description of the business and what we do or the services we provide. This is usually 2-3 sentences of powerful, memorable, and concise text that targets your viewers’ needs.
4. Call to Action
The goal of the website’s homepage is to pique the interest of visitors and prompt them to delve deeper into the pages of our website. A call to action (CTA) is one way to pull people into the interior pages, begin the selling cycle, or at least initiate direct contact. CTA areas or buttons can be linked to contact forms, subscription enrollment forms, or other pages within the website that provide more information. The easier and more intriguing you make it for the visitor to click this CTA, the better the chance they will invest time in browsing the website.
5. Social Proof
Providing testimonials or customer/client reviews is another powerful way to stimulate trust and establish your expertise. The social proof shows new visitors that you know what you’re doing and provides key insights to the product or service. Success stories are a great way to inspire a positive first impression.
6. Photos
To put it simply, people are visual. A photo of someone using the product or engaging with your service is better than a stationary photo of your product because the viewer is able to connect with it. Our brains are hardwired in such a way that our eyes are drawn to human faces and we empathize with the perceived emotion. Photos add a sense of credibility and integrity to a website.
7. Text Content
The language of your site matters. It is best to not only describe what you do but to connect it to why it matters to the visitor. Keep your content simple and easy to read, and avoid industry jargon by being
conscious of your target audience. Connecting with the visitor, using a brand-consistent voice, and convincing them about how they will benefit from you is essential for maintaining interest.
8. Footer
An informational footer is arguably as important as your header navigation. Once the visitor reaches the end of the homepage, this area should provide three features: contact information, links, and social media integration. Contact information encourages the visitor to get in contact with you. Links can be a helpful way to provide a mini-sitemap or encourage the user to check out interior pages. Providing links to social media accounts is a great way to encourage visitors to engage with your company on another medium. It also adds to your social proof and provides another method of connection and communication.